weather/requirements.txt
Flask==1.1.2
line-bot-sdk==1.8.0
pandas==1.1.4
beautifulsoup4==4.9.3
urllib3==1.26.2
requests==2.25.1
lxml==4.6.2